fixing dropdown for mobile devices + add switcher networks
