+++ packages/boba/gateway/src/components/layout/Header/styles.ts
@@ -12,6 +12,7 @@ export const HeaderContainer = styled.div`
   align-items: center;
   justify-content: flex-start;
   backdrop-filter: blur(7.5px);
+  z-index: 9999;
   ${(props) =>
     props.theme.name === 'light' &&
     css`
